requirements.txt
requests
pyinstaller_versionfile
cryptography
auto-py-to-exe
pylint
portablemc[certifi]
customtkinter
pillow